src: create helper for reading Uint32BE
Fixes: https://github.com/nodejs/node/issues/34827
