+++ src/node_sockaddr.cc
@@ -1,5 +1,6 @@
 #include "node_sockaddr-inl.h"  // NOLINT(build/include)
 #include "env-inl.h"
+#include "base64-inl.h"
 #include "base_object-inl.h"
 #include "memory_tracker-inl.h"
 #include "uv.h"
@@ -302,7 +303,7 @@ bool in_network_ipv6_ipv4(
     return false;
 
   ptr += sizeof(mask);
-  uint32_t check = ptr[0] << 24 | ptr[1] << 16 | ptr[2] << 8 | ptr[3];
+  uint32_t check = ReadUint32BE(ptr);
 
   return (check & m) == (htonl(net_in->sin_addr.s_addr) & m);
 }
